backend/requirements.txt
Flask==2.0.1
requests==2.25.1
dateparser==1.0.0
google-api-python-client==2.15.0
google-auth-oauthlib==0.4.5
Authlib==0.15.3
itsdangerous==2.0.1
gunicorn==20.1.0
SQLAlchemy==1.4.20
psycopg2-binary==2.9.1
loguru==0.5.3
Flask-Login==0.5.0
oauthlib==3.1.1
pyOpenSSL==20.0.1
rq==1.9.0
redis==3.5.3
marshmallow==3.13.0
flask-smorest==0.32.0
pandas==1.3.1
msal==1.15.0
XlsxWriter==3.0.2
